Extract type-keyed serializers into `serializers.py`; unify Harness obs channels
Move the pure serializers (`Serializers`, `StatefulSerializer`, `Timestamped`) and
the shared `name + suffix` unfold (`expand_suffixed`) out of `ds_writer_agent` into
`positronic/dataset/serializers.py`, reused by both the writer's `_append` and the
Harness obs assembly. The Harness now treats every observation channel as
`(receiver, serializer-or-None)` — `State` flows through `Serializers.robot_state_obs`,
grip passes through — so there is no robot-vs-grip special-casing and the per-component
split lives in the serializer, not the harness.
